Allow unsimplified constants
For Expression.add and Expression.subtract, allow unsimplified
constants to remain in the expression.

var exp = new Expression(“x”).add(3);
exp = exp.add(2, false);
console.log(exp.toString()); // x + 3 + 2

This required a pretty big refactor, as the expression’s constant was
initially stored as a single Fraction object. This property needed to
be changed to an array of Fraction objects.